requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
filelock==3.12.2
firebase_admin==6.2.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
mysql-connector-python==8.0.33
networkx==3.1
numpy==1.25.0
pipreqs==0.4.13
protobuf==3.20.3
requests==2.31.0
sympy==1.12
numpy==1.24.3
Pyrebase4==4.5.0
Requests==2.31.0
torch==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
yarg==0.1.9